add recording metadata cols instead of dict
