+++ alembic/versions/31e1203361da_add_recording_metadata_cols.py
@@ -1,0 +1,40 @@
+"""add recording metadata cols
+
+Revision ID: 31e1203361da
+Revises: aeda129e5d02
+Create Date: 2026-06-24 10:43:18.022495
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+import sqlmodel
+
+
+# revision identifiers, used by Alembic.
+revision: str = "31e1203361da"
+down_revision: Union[str, None] = "aeda129e5d02"
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column("recording", sa.Column("date_of_recording", sa.Date(), nullable=True))
+    op.add_column("recording", sa.Column("client_date_of_birth", sa.Date(), nullable=True))
+    op.add_column("recording", sa.Column("client_name", sqlmodel.sql.sqltypes.AutoString(), nullable=True))
+    op.add_column("recording", sa.Column("case_id", sqlmodel.sql.sqltypes.AutoString(), nullable=True))
+    op.add_column("recording", sa.Column("subject", sqlmodel.sql.sqltypes.AutoString(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column("recording", "subject")
+    op.drop_column("recording", "case_id")
+    op.drop_column("recording", "client_name")
+    op.drop_column("recording", "client_date_of_birth")
+    op.drop_column("recording", "date_of_recording")
+    # ### end Alembic commands ###
