--- alembic/versions/82bd0836bbb7_add_metadata_column_to_recordings.py
@@ -1,48 +1,0 @@
-"""add metadata column to recordings
-
-Revision ID: 82bd0836bbb7
-Revises: aeda129e5d02
-Create Date: 2026-06-18 14:34:05.165879
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = "82bd0836bbb7"
-down_revision: Union[str, None] = "aeda129e5d02"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "recording",
-        sa.Column(
-            "recording_metadata",
-            postgresql.JSONB(astext_type=sa.Text()),
-            nullable=False,
-            server_default=sa.text("'{}'::jsonb"),  # manual introduction of server default
-        ),
-    )
-
-    # manual edit
-    # backfil existing rows with server default, then remove to sync with model
-    op.alter_column(
-        "recording",
-        "recording_metadata",
-        server_default=None,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("recording", "recording_metadata")
-    # ### end Alembic commands ###
