+++ common/database/postgres_models.py
@@ -1,6 +1,6 @@
-from datetime import UTC, datetime
+from datetime import UTC, date, datetime
 from enum import StrEnum, auto
-from typing import Any, TypedDict
+from typing import TypedDict
 from uuid import UUID, uuid4
 
 from sqlalchemy import TIMESTAMP, Column, Enum, ForeignKey, Text, text
@@ -166,10 +166,12 @@ class Recording(BaseTableMixin, table=True):
     s3_file_key: str
     transcription_id: UUID | None = Field(default=None, foreign_key="transcription.id", ondelete="SET NULL")
     transcription: "Transcription" = Relationship(back_populates="recordings")
-    recording_metadata: dict[str, Any] = Field(
-        default_factory=dict,
-        sa_column=Column(JSONB, nullable=False),
-    )
+    # metadata fields:
+    date_of_recording: date | None = Field(default=None)
+    client_date_of_birth: date | None = Field(default=None)
+    client_name: str | None = Field(default=None)
+    case_id: str | None = Field(default=None)
+    subject: str | None = Field(default=None)
 
 
 class Chat(BaseTableMixin, table=True):
